backend/requirements.txt
python-dotenv
rich
beautifulsoup4
requests
pydantic
